Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/seeking/seek-to-negative-time-expected.txt


FAIL seek to negative time assert_equals: currentTime after setting expected 0 but got -1
